mpin_user: Initial driver commit - pin user buffers to physical addresses
In dpdk/spdk and other user-mode drivers projects. We want a way to lock
memory buffers to physical memory. So we can give them to DMA/RDAM engines
and so on ...
With this driver we can do this.

The API philosophy is:
  You open a special file MPIN_USER_PATH (see mpin_user.h)
  and ADD buffers to pin, on that FD. when FD is closed all added
  buffers will unpin.
  There is also an API for unpinning if buffer needs freeing.
  [API is through calling IOCTLs on the FD]

Very old Kernels, that do not have xarray.h will not implement m-pinning
but the API will be there as a NO-OP, so apps can be agnostic to the Kernel
version.
On those old Kernels, 2M buffers did not move. So it is OK.

Original code is from a posted kernel patch:
	Subject: uacce: Add uacce_ctrl misc device
	From: Zhou Wang <wangzhou1-AT-hisilicon.com>
	See: https://lwn.net/Articles/843432/

[SQUASHMEs]
  a5ab873 mpin: Fix compiler warning when ENABLE_MPIN is OFF
  9b5d673 mpin: Add proper MODULE_VERSION section
  5033a5e mpin: fix the mpin_user.h API file
  fe75d17 SQUASHME: mpin_user: Initial driver commit - header was broken

Signed-off-by: Boaz Harrosh <boaz@weka.io>